refactor: extract classes from main file
