+++ src/main/kotlin/lotto/LottoMessages.kt
@@ -1,0 +1,11 @@
+package lotto
+
+object LottoMessages {
+    val welcome = "Welcome to Lotto Machine! Please, insert the amount do you want to " +
+            "invest on your Lotto Ticket. (It has to be divisible by 1000):"
+    val tryAgainDivisible = " [ERROR] The amount has to be divisible by 1000! Please insert a correct amount value:"
+    val winningNumbers = "Please, Enter 6 winning numbers (coma-separated):"
+    val invalidWinningNumbers = "[ERROR] Invalid input. Please enter 6 winning numbers (coma-separated):"
+    val invalidInput = "[ERROR] Invalid Input. Please try again:"
+    val bonusNumber = "Digit a bonus number:"
+}
